:construction: very basic network settings pane finished
