+++ src/panes/index.jsx
@@ -3,7 +3,7 @@ import ManualInputPane from './ManualInput'
 import TestComponent from './TestComponent'
 import SplitPane from './SplitPane'
 import StatusPane from './StatusPane'
-import WifiPane from './WifiPane'
+// import WifiPane from './WifiPane'
 
 const panes = {
   test: { title: 'Test', Component: TestComponent },
@@ -12,7 +12,7 @@ const panes = {
   status: { title: 'Status', Component: StatusPane },
   vSplit: { title: 'Split Vertically', Component: (props) => (<SplitPane {...props} direction='row' />) },
   hSplit: { title: 'Split Horizontally', Component: (props) => (<SplitPane {...props} />) },
-  wifi: { title: 'Wifi', Component: WifiPane }
+  // wifi: { title: 'Wifi', Component: WifiPane }
 }
 
 export default panes
